+++ src/version2/parameters/deleteUiModification.ts
@@ -1,0 +1,4 @@
+export interface DeleteUiModification {
+  /** The ID of the UI modification. */
+  uiModificationId: string;
+}