+++ src/version2/parameters/search.ts
@@ -5,7 +5,7 @@ export interface Search {
    *
    * `usages` Returns the project and issue types that use the status in their workflow.
    */
-  expand?: 'usages' | ('usages')[] | string | string[];
+  expand?: 'usages' | 'usages'[] | string | string[];
   /** The project the status is part of or null for global statuses. */
   projectId?: string;
   /** The index of the first item to return in a page of results (page offset). */